requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
pyyaml
requests